+++ packages/leancode_lint/test/prefix_widgets_returning_slivers_test.dart
@@ -1,4 +1,3 @@
-import 'package:analyzer/src/lint/registry.dart';
 import 'package:analyzer_testing/analysis_rule/analysis_rule.dart';
 import 'package:leancode_lint/lints/prefix_widgets_returning_slivers.dart';
 import 'package:test_reflective_loader/test_reflective_loader.dart';
@@ -13,19 +12,14 @@ void main() {
 
 @reflectiveTest
 class PrefixWidgetsReturningSliversTest extends AnalysisRuleTest {
-  final rule = PrefixWidgetsReturningSlivers();
-
   @override
   void setUp() {
-    Registry.ruleRegistry.registerWarningRule(rule);
+    rule = PrefixWidgetsReturningSlivers();
     super.setUp();
 
     addMocks([MockLibrary.flutter]);
     addAnalysisOptions();
   }
-
-  @override
-  String get analysisRule => rule.name;
 
   Future<void> test_return_from_internal_blocks() async {
     await assertDiagnostics(