+++ packages/leancode_lint/test/start_comments_with_space_test.dart
@@ -1,4 +1,3 @@
-import 'package:analyzer/src/lint/registry.dart';
 import 'package:analyzer_testing/analysis_rule/analysis_rule.dart';
 import 'package:leancode_lint/lints/start_comments_with_space.dart';
 import 'package:test_reflective_loader/test_reflective_loader.dart';
@@ -11,16 +10,11 @@ void main() {
 
 @reflectiveTest
 class StartCommentsWithSpaceTest extends AnalysisRuleTest {
-  final rule = StartCommentsWithSpace();
-
   @override
   void setUp() {
-    Registry.ruleRegistry.registerWarningRule(rule);
+    rule = StartCommentsWithSpace();
     super.setUp();
   }
-
-  @override
-  String get analysisRule => rule.name;
 
   Future<void> test_main() async {
     await assertDiagnostics(